feat: use tmp & clean up
