+++ loader/palera1nLoader/UI/Views/JsonVC.swift
@@ -21,174 +21,6 @@ class JsonVC: UIViewController, UITableViewDelegate, UITableViewDataSource {
     var errorMessage = "Unable to fetch bootstraps."
     var tableView: UITableView!
     
-    func installDebFile(file: String) {
-        UIApplication.shared.isIdleTimerDisabled = true
-        let title: String.LocalizationValue = "\(local("DOWNLOADING"))"
-        let downloadAlert = UIAlertController.downloading(title)
-        present(downloadAlert, animated: true)
-        
-        let downloadUrl = getManagerURL(envInfo.jsonInfo!, file)!
-
-        downloadFile(url: URL(string: downloadUrl)!, forceBar: true, completion:{(path:String?, error:Error?) in
-            DispatchQueue.main.async {
-                downloadAlert.dismiss(animated: true) {
-                    if (error == nil) {
-                        let installingAlert = UIAlertController.spinnerAlert("INSTALLING")
-                        self.present(installingAlert, animated: true) {
-                            bootstrap.installDebian(deb: path!, completion:{(msg:String?, error:Int?) in
-                                installingAlert.dismiss(animated: true) {
-                                    if (error == 0) {
-                                        let alert = UIAlertController.error(title: local("DONE_INSTALL"), message: local("DONE_INSTALL_SUB"))
-                                        self.present(alert, animated: true)
-                                    } else {
-                                        let alert = UIAlertController.error(title: local("ERROR_INSTALL"), message: "Status: \(errorString(Int32(error!)))")
-                                        self.present(alert, animated: true)
-                                    }
-                                }
-                            })
-                        }
-                    } else {
-                        let alert = UIAlertController.error(title: "Download Failed", message: error!.localizedDescription)
-                        self.present(alert, animated: true)
-                    }
-                }
-            }
-        })
-    }
-    
-    func downloadFile(url: URL, forceBar: Bool = false, output: String? = nil, completion: @escaping (String?, Error?) -> Void) {
-        let tempDir = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first!
-        var destinationUrl = tempDir.appendingPathComponent(url.lastPathComponent)
-        if (output != nil) {destinationUrl = tempDir.appendingPathComponent(output!)}
-
-        let session = URLSession(configuration: URLSessionConfiguration.default, delegate: nil, delegateQueue: nil)
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        
-        let task = session.dataTask(with: request, completionHandler: { data, response, error in
-            if error == nil {
-                if let response = response as? HTTPURLResponse {
-                    if response.statusCode == 200 {
-                        if let data = data {
-                            if let _ = try? data.write(to: destinationUrl, options: Data.WritingOptions.atomic) {
-                                spawn(command: "/cores/binpack/bin/mv", args: [destinationUrl.path, "/var/mobile/Library/palera1n/downloads/\(destinationUrl.lastPathComponent)"])
-                                completion("/var/mobile/Library/palera1n/downloads/\(destinationUrl.lastPathComponent)", error)
-                                log(type: .info, msg: "Saved to: /var/mobile/Library/palera1n/downloads/\(destinationUrl.lastPathComponent)")
-                            } else {
-                                completion(destinationUrl.path, error)
-                                log(type: .error, msg: "Failed to save file at: \(destinationUrl.path)")
-                            }
-                        } else {
-                            completion(destinationUrl.path, error)
-                            log(type: .error, msg: "Failed to download: \(request)")
-                        }
-                    } else {
-                        completion(destinationUrl.path, error)
-                        log(type: .error, msg: "Unknown error on download: \(response.statusCode) - \(request)")
-                    }
-                }
-            } else {
-                completion(destinationUrl.path, error)
-                log(type: .error, msg: "Failed to download: \(request)")
-            }
-        })
-        
-        if (url.pathExtension == "zst" || url.pathExtension == "tar" || forceBar) {
-            observation = task.progress.observe(\.fractionCompleted) { progress, _ in
-                DispatchQueue.main.async {
-                    progressDownload.setProgress(Float(progress.fractionCompleted/1.0), animated: true)
-                }
-            }
-        }
-        task.resume()
-    }
-    
-    func installStrap(file: String, completion: @escaping () -> Void) {
-        UIApplication.shared.isIdleTimerDisabled = true
-        let downloadAlert = UIAlertController.downloading("DL_STRAP")
-        present(downloadAlert, animated: true)
-
-        let bootstrapUrl = getBootstrapURL(envInfo.jsonInfo!)!
-        let pkgmgrUrl = getManagerURL(envInfo.jsonInfo!, file)!
-        
-        downloadFile(url: URL(string: pkgmgrUrl)!, completion:{(path:String?, error:Error?) in
-            if (error != nil) {
-                DispatchQueue.main.async {
-                    downloadAlert.dismiss(animated: true) {
-                        let alert = UIAlertController.error(title: local("DOWNLOAD_FAIL"), message: error.debugDescription)
-                        self.present(alert, animated: true)
-                    }
-                }
-            }
-        })
-
-        self.downloadFile(url:  URL(string: bootstrapUrl)!, completion:{(path:String?, error:Error?) in
-            DispatchQueue.main.async {
-                downloadAlert.dismiss(animated: true) {
-                    if (error == nil) {
-                        let installingAlert = UIAlertController.spinnerAlert("INSTALLING")
-                        self.present(installingAlert, animated: true) {
-                            bootstrap.installBootstrap(tar: path!, deb: "\(file).deb", completion:{(msg:String?, error:Int?) in
-                                installingAlert.dismiss(animated: true) {
-                                    if (error == 0) {
-                                        let message = local("PASSWORD")
-                                        let alertController = UIAlertController(title: local("PASSWORD_SET"), message: message, preferredStyle: .alert)
-                                        alertController.addTextField() { (password) in
-                                            password.placeholder = local("PASSWORD_TEXT")
-                                            password.isSecureTextEntry = true
-                                            password.keyboardType = UIKeyboardType.asciiCapable
-                                        }
-
-                                        alertController.addTextField() { (repeatPassword) in
-                                            repeatPassword.placeholder = local("PASSWORD_REPEAT")
-                                            repeatPassword.isSecureTextEntry = true
-                                            repeatPassword.keyboardType = UIKeyboardType.asciiCapable
-                                        }
-
-                                        let setPassword = UIAlertAction(title: local("SET"), style: .default) { _ in
-                                            helper(args: ["-P", alertController.textFields![0].text!])
-                        
-                                            alertController.dismiss(animated: true) {
-                                                let alert = UIAlertController.error(title: local("DONE_INSTALL"), message: local("DONE_INSTALL_SUB"))
-                                                self.present(alert, animated: true)
-                                                completion()
-                                            }
-                                        }
-                                        setPassword.isEnabled = false
-                                        alertController.addAction(setPassword)
-
-                                        NotificationCenter.default.addObserver(
-                                            forName: UITextField.textDidChangeNotification,
-                                            object: nil,
-                                            queue: .main
-                                        ) { notification in
-                                            let passOne = alertController.textFields![0].text
-                                            let passTwo = alertController.textFields![1].text
-                                            if (passOne!.count > 253 || passOne!.count > 253) {
-                                                setPassword.setValue(local("TOO_LONG"), forKeyPath: "title")
-                                            } else {
-                                                setPassword.setValue(local("SET"), forKeyPath: "title")
-                                                setPassword.isEnabled = (passOne == passTwo) && !passOne!.isEmpty && !passTwo!.isEmpty
-                                            }
-                                        }
-                                        self.present(alertController, animated: true)
-                                    } else {
-                                        let errStr = String(cString: strerror(Int32(error!)))
-                                        let alert = UIAlertController.error(title: local("ERROR_INSTALL"), message: errStr)
-                                        self.present(alert, animated: true)
-                                    }
-                                }
-                            })
-                        }
-                    } else {
-                        let alert = UIAlertController.error(title: local("DOWNLOAD_FAIL"), message: error.debugDescription)
-                        self.present(alert, animated: true)
-                    }
-                }
-            }
-        })
-    }
-    
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         envInfo.nav = navigationController!
